refactor: move requirement module and work views to view.py
